Project Structure modified - to a better modular approach - coverage implemented (18%)
