+++ project/home/templates/home.html
@@ -8,15 +8,15 @@
 
     <div class="container">
 
-      <a href="{{ url_for('create_politician') }}" class="btn btn-default btn-lg" role="button">Create Politician</a>
+      <a href="{{ url_for('politicians.create_politician') }}" class="btn btn-default btn-lg" role="button">Create Politician</a>
       <h3>Politics:</h3>
       {% for politic in politics %}
         <strong>Public Name: </strong>{{politic.publicName}} <br>
         <strong>Complete Name: </strong>{{politic.completeName}} <br>
 
-        <a href="{{ url_for('politician', idPolitician=politic.idPolitician) }}"  class="btn btn-default btn-sm" role="button"> See Politician </a>
+        <a href="{{ url_for('politicians.politician', idPolitician=politic.idPolitician) }}"  class="btn btn-default btn-sm" role="button"> See Politician </a>
 
-        <a href="{{ url_for('edit_politician', idPol=politic.idPolitician) }}" class="btn btn-default btn-sm" role="button">Edit</a>
+        <a href="{{ url_for('politicians.edit_politician', idPol=politic.idPolitician) }}" class="btn btn-default btn-sm" role="button">Edit</a>
 
 
         <br>
