+++ project/home/templates/index.html
@@ -6,9 +6,9 @@
       
       <div class="section-content">
         <h2>Welcome to Polex.ORG</h2>
-        <a href="{{ url_for('signup') }}" class="btn-primary">Sign up</a>
-        <a href="{{ url_for('about') }}" class="btn-secondary">Learn more</a>
-        <a href="{{ url_for('politician') }}" class="btn-secondary">See Politicians</a>
+        <a href="{{ url_for('users.signup') }}" class="btn-primary">Sign up</a>
+        <a href="{{ url_for('home.about') }}" class="btn-secondary">Learn more</a>
+        <a href="{{ url_for('politicians.politician') }}" class="btn-secondary">See Politicians</a>
  
       </div>
