+++ project/home/templates/welcome.html
