+++ project/static/css/bootstrap-theme.css
