+++ project/static/fonts/glyphicons-halflings-regular.woff2
