+++ project/users/templates/login.html
@@ -1,0 +1,40 @@
+{% extends "base.html" %}
+
+{% block content %}
+
+
+<main class="container signup-section">
+  <div class="section-content">
+  <h1>Please login</h1>
+    <form class="form-signin" role="form" method="post" action="">
+      {{ form.csrf_token }}
+      <p>
+        <div class="form-group">
+          {{ form.email(placeholder="email") }}
+          <span class="error">
+            {% if form.email.errors %}
+            {% for error in form.email.errors %}
+            {{ error }}
+            {% endfor %}
+            {% endif %}
+          </span>
+        </div>
+      </p>
+      <p>
+        <div class="form-group">
+          {{ form.password(placeholder="password") }}
+          <span class="error">
+            {% if form.password.errors %}
+            {% for error in form.password.errors %}
+            {{ error }}
+            {% endfor %}
+            {% endif %}
+          </span>
+        </div>
+      </p>
+      {{ form.submit(class="btn-primary") }}
+    </form>
+
+  </div>
+</main>
+{% endblock %}
