+++ run.py
@@ -1,0 +1,5 @@
+from project import app
+
+
+if __name__ == '__main__':
+    app.run()
